+++ apps/fluent-tester/src/E2E/Button/pages/ButtonPageObject.ts
@@ -17,33 +17,28 @@ class ButtonPageObject extends BasePage {
   /******************************************************************/
   /**************** UI Element Interaction Methods ******************/
   /******************************************************************/
-  didOnClickCallbackFire(): boolean {
-    const callbackText = By(BUTTON_ON_PRESS_DEPRECATED);
-    browser.waitUntil(
-      () => {
-        return callbackText.isDisplayed();
-      },
-      {
-        timeout: this.waitForPageTimeout,
-        timeoutMsg: 'The OnClick callback did not fire.',
-        interval: 1000,
-      },
-    );
+  async didOnClickCallbackFire(): Promise<boolean> {
+    const callbackText = await By(BUTTON_ON_PRESS_DEPRECATED);
+    await browser.waitUntil(async () => await callbackText.isDisplayed(), {
+      timeout: this.waitForPageTimeout,
+      timeoutMsg: 'The OnClick callback did not fire.',
+      interval: 1000,
+    });
 
-    return callbackText.isDisplayed();
+    return await callbackText.isDisplayed();
   }
 
   /* Sends a Keyboarding command on a specific UI element */
-  sendKey(buttonSelector: ButtonSelector, key: string): void {
-    this.getButtonSelector(buttonSelector).addValue(key);
+  async sendKey(buttonSelector: ButtonSelector, key: string): Promise<void> {
+    await (await this.getButtonSelector(buttonSelector)).addValue(key);
   }
 
   /* Returns the correct WebDriverIO element from the Button Selector */
-  getButtonSelector(buttonSelector?: ButtonSelector): WebdriverIO.Element {
+  async getButtonSelector(buttonSelector?: ButtonSelector): Promise<WebdriverIO.Element> {
     if (buttonSelector == ButtonSelector.PrimaryButton) {
-      return this._primaryComponent;
+      return await this._primaryComponent;
     }
-    return this._primaryComponent;
+    return await this._primaryComponent;
   }
 
   /*****************************************/