+++ apps/fluent-tester/src/E2E/Button/specs/Button.spec.macos.ts
@@ -4,20 +4,20 @@ import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('Button Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy();
   });
 
-  it('Click and navigate to Button test page', () => {
+  it('Click and navigate to Button test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    ButtonPageObject.scrollToComponentButton();
-    ButtonPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+    await ButtonPageObject.scrollToComponentButton();
+    await ButtonPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToButtonPage();
-    ButtonPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToButtonPage();
+    await ButtonPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(ButtonPageObject.isPageLoaded()).toBeTruthy();
+    await expect(await ButtonPageObject.isPageLoaded()).toBeTruthy();
   });
 });