+++ apps/fluent-tester/src/E2E/Svg/specs/Svg.spec.macos.ts
@@ -4,20 +4,20 @@ import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('Svg Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy();
   });
 
-  it('Click and navigate to Svg test page', () => {
+  it('Click and navigate to Svg test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    SvgPageObject.scrollToComponentButton();
-    SvgPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+    await SvgPageObject.scrollToComponentButton();
+    await SvgPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToSvgPage();
-    SvgPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToSvgPage();
+    await SvgPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(SvgPageObject.isPageLoaded()).toBeTruthy();
+    await expect(await SvgPageObject.isPageLoaded()).toBeTruthy();
   });
 });