+++ apps/fluent-tester/src/E2E/common/BasePage.macos.ts
@@ -3,7 +3,7 @@ export const COMPONENT_SCROLL_COORDINATES = { x: -0, y: -100 }; // These are the
 
 /* Mac-Specific Selector. We use this to get elements on the test page.
  * Specifically, this is the xpath selector. See https://v6.webdriver.io/docs/selectors.html#xpath */
-export function By(identifier: string): WebdriverIO.Element {
+export function By(identifier: string) {
   return $('//*[@identifier="' + identifier + '"]');
 }
 
@@ -26,100 +26,84 @@ export class BasePage {
   /******************************************************************/
   /**************** UI Element Interaction Methods ******************/
   /******************************************************************/
-  getAccessibilityRole(): string {
-    return this._primaryComponent.getAttribute('ControlType');
+  async getAccessibilityRole(): Promise<string> {
+    return await this._primaryComponent.getAttribute('ControlType');
   }
 
   /* Gets the accessibility label of an UI element given the selector */
-  getAccessibilityLabel(componentSelector: ComponentSelector): string {
+  async getAccessibilityLabel(componentSelector: ComponentSelector): Promise<string> {
     switch (componentSelector) {
       case ComponentSelector.Primary:
-        return this._primaryComponent.getAttribute('Name');
+        return await this._primaryComponent.getAttribute('Name');
 
       case ComponentSelector.Secondary:
-        return this._secondaryComponent.getAttribute('Name');
+        return await this._secondaryComponent.getAttribute('Name');
     }
   }
 
   /* Returns true if the test page has loaded. To determine if it's loaded, each test page has a specific UI element we attempt to locate.
    * If this UI element is located, we know the page as loaded correctly. The UI element we look for is a Text component that contains
    * the title of the page (this._testPage returns that UI element)  */
-  isPageLoaded(): boolean {
-    return this._testPage.isDisplayed();
+  async isPageLoaded(): Promise<boolean> {
+    return await this._testPage.isDisplayed();
   }
 
   /* Returns true if the test page's button is displayed (the button that navigates to each test page) */
-  isButtonInView(): boolean {
-    return this._pageButton.isDisplayed();
+  async isButtonInView(): Promise<boolean> {
+    return await this._pageButton.isDisplayed();
+  }
+
+  async clickComponent(): Promise<void> {
+    await this._primaryComponent.click();
   }
 
   /* Scrolls until the desired test page's button is displayed. We use the scroll viewer UI element as the point to start scrolling.
    * We use a negative number as the Y-coordinate because that enables us to scroll downwards */
-  scrollToComponentButton(): void {
-    /* TODO: Implement (not needed yet) */
+  async scrollToComponentButton(): Promise<void> {
+    /* Not Needed Yet */
   }
 
   /* Waits for the test page to load. If the test page doesn't load before the timeout, it causes the test to fail. */
-  waitForPageDisplayed(timeout?: number): void {
-    browser.waitUntil(
-      () => {
-        return this.isPageLoaded();
-      },
-      {
-        timeout: timeout ?? this.waitForPageTimeout,
-        timeoutMsg: this._pageName + ' did not render correctly. Please see /errorShots for more information.',
-        interval: 1000,
-      },
-    );
+  async waitForPageDisplayed(timeout?: number): Promise<void> {
+    await browser.waitUntil(async () => await this.isPageLoaded(), {
+      timeout: timeout ?? this.waitForPageTimeout,
+      timeoutMsg: this._pageName + ' did not render correctly. Please see /errorShots for more information.',
+      interval: 1500,
+    });
   }
 
   /* Waits for the test page's button to be displayed. If the button doesn't load before the timeout, it causes the test to fail. */
-  waitForButtonDisplayed(timeout?: number): void {
-    browser.waitUntil(
-      () => {
-        return this.isButtonInView();
-      },
-      {
-        timeout: timeout ?? this.waitForPageTimeout,
-        timeoutMsg: 'Could not find the button to navigate to ' + this._pageName + '. Please see /errorShots for more information.',
-        interval: 1000,
-      },
-    );
+  async waitForButtonDisplayed(timeout?: number): Promise<void> {
+    await browser.waitUntil(async () => await this.isButtonInView(), {
+      timeout: timeout ?? this.waitForPageTimeout,
+      timeoutMsg: 'Could not find the button to navigate to ' + this._pageName + '. Please see /errorShots for more information.',
+      interval: 1500,
+    });
   }
 
   /* Waits for the primary UI test element to be displayed. If the element doesn't load before the timeout, it causes the test to fail. */
-  waitForPrimaryElementDisplayed(timeout?: number): void {
-    browser.waitUntil(
-      () => {
-        return this._primaryComponent.isDisplayed();
-      },
-      {
-        timeout: timeout ?? this.waitForPageTimeout,
-        timeoutMsg:
-          'The primary UI element for testing did not display correctly. Please see /errorShots of the first failed test for more information.',
-        interval: 1000,
-      },
-    );
+  async waitForPrimaryElementDisplayed(timeout?: number): Promise<void> {
+    await browser.waitUntil(async () => await this._primaryComponent.isDisplayed(), {
+      timeout: timeout ?? this.waitForPageTimeout,
+      timeoutMsg:
+        'The primary UI element for testing did not display correctly. Please see /errorShots of the first failed test for more information.',
+      interval: 1500,
+    });
   }
 
   /* Scrolls to the primary UI test element until it is displayed. It uses the ScrollView that encapsulates each test page. */
-  scrollToTestElement(): void {
-    /* TODO: Implement (not needed yet) */
+  async scrollToTestElement(): Promise<void> {
+    /* Not Needed Yet */
   }
 
   /* A method that allows the caller to pass in a condition. A wrapper for waitUntil(). Once testing becomes more extensive,
    * this will allow cleaner code within all the Page Objects. */
-  waitForCondition(condition?: () => boolean, errorMsg?: string, timeout?: number): void {
-    browser.waitUntil(
-      () => {
-        return condition();
-      },
-      {
-        timeout: timeout ?? this.waitForPageTimeout,
-        timeoutMsg: errorMsg ?? 'Error. Please see /errorShots and logs for more information.',
-        interval: 1000,
-      },
-    );
+  async waitForCondition(condition?: () => Promise<boolean>, errorMsg?: string, timeout?: number): Promise<void> {
+    await browser.waitUntil(async () => await condition(), {
+      timeout: timeout ?? this.waitForPageTimeout,
+      timeoutMsg: errorMsg ?? 'Error. Please see /errorShots and logs for more information.',
+      interval: 1000,
+    });
   }
 
   /*****************************************/
@@ -134,26 +118,26 @@ export class BasePage {
 
   // Returns: UI Element
   // The Text component on each test page containing the title of that page. We can use this to determine if a test page has loaded correctly.
-  get _testPage(): WebdriverIO.Element {
+  get _testPage() {
     return By(DUMMY_CHAR);
   }
 
   // Returns: UI Element
   // The primary UI element used for testing on the given test page.
-  get _primaryComponent(): WebdriverIO.Element {
+  get _primaryComponent() {
     return By(DUMMY_CHAR);
   }
 
   // Returns: UI Element
   // The secondary UI element used for testing on the given test page. Often times, we'll want to set a
   // prop on one component, and not set it on another to verify certain behaviors. This is why we have this secondary component.
-  get _secondaryComponent(): WebdriverIO.Element {
+  get _secondaryComponent() {
     return By(DUMMY_CHAR);
   }
 
   // Returns: UI Element
   // The button that navigates you to the component's test page.
-  get _pageButton(): WebdriverIO.Element {
+  get _pageButton() {
     return By(DUMMY_CHAR);
   }