+++ packages/theming/android-theme/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/android-theme
 
-This log was last generated on Thu, 14 Jul 2022 18:09:50 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Aug 2022 23:21:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.0
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Minor changes
+
+- Added unit tests (nkhalil942@gmail.com)
 
 ## 0.11.0